requirements.txt
APScheduler==3.10.1
asgiref==3.5.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-apscheduler==0.6.2
django-cors-headers==3.11.0
django-paypal==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
pandas==2.0.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
